Avoid circular imports workaround by moving  TemplateElement in pydoctor.templatewriter
Cleanup commented HTML.

Cleanup the __init__ methods of class pages to only accept what's necessary.
