+++ pydoctor/templates/common.html
@@ -1,12 +1,6 @@
 <!DOCTYPE html>
 <html xmlns:t="http://twistedmatrix.com/ns/twisted.web.template/0.1" t:render="all">
   <meta name="pydoctor-template-version" content="1" />
-  
-  <!-- <head>
-    <title><t:slot name="title">Something</t:slot> : API documentation</title>
-    
-    <t:transparent t:render="meta">Meta</t:transparent>
-  </head> -->
 
   <t:transparent t:render="head">Head</t:transparent>
