+++ pydoctor/templates/summary.html
@@ -1,11 +1,6 @@
 <!DOCTYPE html>
 <html xmlns:t="http://twistedmatrix.com/ns/twisted.web.template/0.1">
   <meta name="pydoctor-template-version" content="1" />
-
-  <!-- <head>
-    <title t:render="title">Something</title>
-    <t:transparent t:render="meta">Meta</t:transparent>
-  </head> -->
   
   <t:transparent t:render="head">Head</t:transparent>
